Redesign main page and removed about
