+++ web-user/src/app/app.component.html
@@ -119,12 +119,6 @@
       </a>
       </mat-list-item>
       <mat-divider></mat-divider>
-      <mat-list-item>
-        <a mat-button class="main-left-menu-item" routerLink="/" l10nTranslate> pageWelcome </a>
-      </mat-list-item>
-      <mat-list-item>
-        <a mat-button class="main-left-menu-item" routerLink="/about" l10nTranslate> pageAbout </a>
-      </mat-list-item>
       <!-- mat-list-item>
         <a mat-button class="main-left-menu-item" routerLink="/chat" l10nTranslate> chatPage </a>
       </mat-list-item -->
