--- web-user/src/app/pages/about/about.component.html
@@ -1,101 +1,0 @@
-<h1 l10nTranslate>pageAbout</h1>
-
-<div *ngIf="errorMessage != null" class="alert alert-danger" color="#ff0000"> {{ errorMessage }} </div>
-
-<div class="card">
-	<div class="card-header">About</div>
-	<div class="card-body">
-		<strong>FreeHackQuest</strong> - <div class="d-inline-block">This is an open source platform for competitions in computer security</div>
-    <hr>
-    <a class="btn btn-link" routerLink="/donate" l10nTranslate>pageDonate</a>
-    <hr>
-    <iframe style="vertical-align: middle; padding-left: 20px" frameborder="0" width="100" height="20" src="https://ghbtns.com/github-btn.html?user=freehackquest&amp;repo=fhq-server&amp;type=star&amp;count=true&amp;size=small"></iframe>
-	</div>
-</div><br>
-<div class="card">
-	<div class="card-header">Statistics</div>
-	<div class="card-body">
-		<p><label>Quests:</label> {{ statisticsCountQuests }}</p>
-		<p><label>All attempts:</label> {{ statisticsAllAttempts }} </p>
-		<p><label>Already solved:</label> {{ statisticsAlreadySolved }} </p>
-		<p><label>Users online:</label> {{ statisticsUsersOnline }} </p>
-	</div>
-</div>
-<br>
-
-<div class="card">
-	<div class="card-header">Playing with us</div>
-	<div class="card-body" id="statistics_playing_with_us">
-    <div class="d-inline-block" *ngFor="let c of cities">
-      | {{ c.city }} ( {{ c.cnt }}) |
-    </div>
-    <br><br><a class="btn btn-info" routerLink="/map-activity">On Map</a>
-  </div>
-</div><br>
-<div class="card">
-  <div class="card-header">Top 10</div>
-  <div class="card-body">
-    <div *ngFor="let w of winners">
-      <strong>{{ w.place }}</strong>
-      [{{ w.rating }} points]
-      <strong>{{ w.university != '' ? '[' + w.university + ']' : '' }} {{ w.nick }}</strong>
-    </div>
-  </div>
-</div>
-
-<app-block-contact-us></app-block-contact-us>
-
-<div class="card">
-  <div class="card-header">License</div>
-  <div class="card-body">
-The MIT License (MIT)<br>
-<br>
-Copyright (c) 2011-2019 FreeHackQuest Team<br>
-<br>
-Permission is hereby granted, free of charge, to any person obtaining a copy of<br>
-this software and associated documentation files (the "Software"), to deal in<br>
-the Software without restriction, including without limitation the rights to<br>
-use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of<br>
-the Software, and to permit persons to whom the Software is furnished to do so,<br>
-subject to the following conditions:<br>
-<br>
-The above copyright notice and this permission notice shall be included in all<br>
-copies or substantial portions of the Software.<br>
-<br>
-THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR<br>
-IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS<br>
-FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR<br>
-COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER<br>
-IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN<br>
-CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.<br>
-  </div>
-</div>
-<br>
-
-<div class="card">
-  <div class="card-header">Distribution</div>
-  <div class="card-body">
-    <h3>Install fhq-server from docker</h3>
-    <p>Docker hub: <a href="https://hub.docker.com/r/sea5kg/fhq-server" target="_blank">https://hub.docker.com/r/sea5kg/fhq-server</a></p>
-    <p>Docker-compose example: <a href="https://github.com/freehackquest/fhq-server/tree/dev/contrib/docker_compose_example">docker_compose_example</a></p>
-    <pre>$ docker pull freehackquest/fhq-server</pre>
-    <hr>
-    <h3>Source code</h3>
-    <p>fhq-server: <a href="http://github.com/freehackquest/fhq-server/" target="_blank">http://github.com/freehackquest/fhq-server</a></p>
-    <h3>Install fhq-server from repository (deprecated)</h3>
-    Ubuntu repo: <a href="https://launchpad.net/~freehackquest/+archive/ubuntu/fhq-server/+packages" target="_blank">https://launchpad.net/~freehackquest/+archive/ubuntu/fhq-server/+packages</a>
-    <pre>$ sudo add-apt-repository ppa:freehackquest/fhq-server<br>$ sudo apt-get update<br>$ sudo apt install fhq-server</pre>
-    <hr>
-  </div>
-</div>
-
-<br>
-<div class="card">
-  <div class="card-header">Developers and designers</div>
-  <div class="card-body">
-    <div *ngFor="let w of developers">
-      {{ w }}
-    </div>
-  </div>
-
-</div><br>
