+++ web-user/src/app/pages/page-not-found/page-not-found.component.html
@@ -4,6 +4,7 @@
 <ul>
   <li><a routerLink="/quests" l10nTranslate>quests</a></li>
   <li><a routerLink="/events" l10nTranslate>events</a></li>
-  <li><a routerLink="/knowledge-base" l10nTranslate>knowledgeBase</a></li>
-  <li><a routerLink="/chat" l10nTranslate>chatPage</a></li>
+  <li><a routerLink="/scoreboard" l10nTranslate>scoreboard</a></li>
+  <!-- li><a routerLink="/knowledge-base" l10nTranslate>knowledgeBase</a></li>
+  <li><a routerLink="/chat" l10nTranslate>chatPage</a></li -->
 </ul>
